Branch BaseClosableWindow 	- RenameWindowViewModel.cs 		- Removed unused namespaces 	- MainWindowViewModel.cs 		- Added view model parameter to RenameWindow 		- Added view model parameter to FormatWindow 	- BaseClosableWindow.cs 		- Inital class setup 			- Constructor to set a DataContext 			- Loaded method to setup RequestClose lambda 	- FormatWindow.xaml 		- Changed Window element to BaseClosableWindow 		- Changed Class to Views 		- Changed xmlns:local to Views 	- FormatWindow.xaml.cs 		- Changed namespace from ViewModels to Views 			- Not sure why it was in ViewModels 		- Changed inheritence from Window to BaseClosableWindow 		- Removed Loaded lambda 	- MainWindow.xaml 		- Changed Window element to BaseClosableWindow 		- Added xmlns:local 	- MainWindow.xaml.cs 		- Changed inheritence from Window to BaseClosableWindow 		- Removed Loaded lambda 	- RenameWindow.xaml 		- Changed Window element to BaseClosableWindow 	- RenameWindow.xaml.cs 		- Changed inheritence from Window to BaseClosableWindow 		- Removed Loaded lambda 	- WaitWindow.xaml.cs 		- Removed unused namespaces
