+++ ViewModels/RenameWindowViewModel.cs
@@ -1,5 +1,4 @@
 ﻿
-using DiskpartGUI.Commands;
 using DiskpartGUI.Helpers;
 using DiskpartGUI.Models;
 using DiskpartGUI.Processes;
