+++ Views/FormatWindow.xaml
@@ -1,9 +1,10 @@
-﻿<Window x:Class="DiskpartGUI.ViewModels.FormatWindow"
+﻿<local:BaseClosableWindow 
+        x:Class="DiskpartGUI.Views.FormatWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:DiskpartGUI.ViewModels"
+        xmlns:local="clr-namespace:DiskpartGUI.Views"
         mc:Ignorable="d"
         Title="{Binding Title}"
         SizeToContent="WidthAndHeight"
@@ -77,4 +78,4 @@
 
 
     </Grid>
-</Window>
+</local:BaseClosableWindow>
