+++ Views/MainWindow.xaml
@@ -1,8 +1,10 @@
-﻿<Window x:Class="DiskpartGUI.Views.MainWindow"
+﻿<local:BaseClosableWindow 
+        x:Class="DiskpartGUI.Views.MainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:DiskpartGUI.Views"
         mc:Ignorable="d"
         Title="Distpart GUI" 
         SizeToContent="WidthAndHeight"
@@ -131,4 +133,4 @@
 
         </DockPanel>
     </Grid>
-</Window>
+</local:BaseClosableWindow>
