+++ Views/MainWindow.xaml.cs
@@ -1,29 +1,15 @@
-﻿using DiskpartGUI.Interfaces;
-using DiskpartGUI.ViewModels;
-using System;
-using System.Windows;
+﻿using DiskpartGUI.ViewModels;
 
 namespace DiskpartGUI.Views
 {
     /// <summary>
     /// Interaction logic for MainWindow.xaml
     /// </summary>
-    public partial class MainWindow : Window
+    public partial class MainWindow : BaseClosableWindow
     {
-        public MainWindow()
+        public MainWindow() : base(new MainWindowViewModel())
         {
             InitializeComponent();
-            Loaded += Window_Loaded;
-        }
-
-        private void Window_Loaded(object sender, RoutedEventArgs e)
-        {
-            DataContext = new MainWindowViewModel();
-
-            if (DataContext is IClosable)
-            {
-                (DataContext as IClosable).RequestClose += (_, __) => this.Close();
-            }
         }
     }
 }
