+++ Views/RenameWindow.xaml
@@ -1,4 +1,5 @@
-﻿<Window x:Class="DiskpartGUI.Views.RenameWindow"
+﻿<local:BaseClosableWindow
+        x:Class="DiskpartGUI.Views.RenameWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -18,13 +19,13 @@
             <RowDefinition Height="Auto"/>
             <RowDefinition Height="Auto"/>
         </Grid.RowDefinitions>
-        
+
         <!-- New Label Text -->
         <Label Content="{Binding NewLabelText, UpdateSourceTrigger=PropertyChanged}" Margin="10" Grid.Row="0" Grid.Column="0" Grid.ColumnSpan="2" HorizontalAlignment="Center" VerticalAlignment="Center"/>
-        
+
         <!-- Label Input Text Box -->
         <TextBox Name="TextBoxLabel" Text="{Binding TextBoxText, Mode=TwoWay, UpdateSourceTrigger=PropertyChanged}" Height="Auto" Width="220" Margin="10" Grid.Row="1" Grid.Column="0"/>
-        
+
         <!-- Apply/Cancel Buttons -->
         <Grid Grid.Row="2" Grid.Column="0">
             <Grid.ColumnDefinitions>
@@ -34,7 +35,7 @@
             <Grid.RowDefinitions>
                 <RowDefinition Height="Auto"/>
             </Grid.RowDefinitions>
-            
+
             <!-- Apply Button -->
             <Button Name="ButtonApply" Content="Apply" Width="100" Height="Auto" Margin="10" Command="{Binding ApplyCommand, UpdateSourceTrigger=PropertyChanged}" Grid.Row="0" Grid.Column="0" VerticalAlignment="Center" HorizontalAlignment="Center"/>
 
@@ -44,4 +45,4 @@
         </Grid>
 
     </Grid>
-</Window>
+</local:BaseClosableWindow>
